requirements.txt
django==1.9.6
psycopg2==2.6.1
djangorestframework==3.3.3
django-filter==0.13.0
django-crispy-forms==1.6.0
django-angular==0.8.2
django-bower==5.1.0
python-dateutil==2.5.3
django-redis==4.4.4
itsdangerous==0.24
celery==3.1.23
requests==2.11.1
